Action: Deleted the Doc and Developer directories
Reason:
The initial version of the CppAD subversion repository had the Doc and 
Developer directories in it. These directories are not source code but are
rather created from the source using OMhelp. They are in the distribution
of CppAD because the user of CppAD is not expected to have OMhelp installed.


git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@2 0e562018-b8fb-0310-a933-ecf073c7c197