configure: Previous commit removed configure script because it is created by automake + autoconf.
Build:
I am used to cvs and forgot that subversion would also commit other changes.
One change was that SetNoExecute was removed from build becasue the subversion
reposiotry had the proper execute settings on the files. 

configure.ac and Doc.omh:
Another change was the updating of the date on the configure.ac and Doc.omh 
files. These change automatically every day and there is no reason to 
store the change in the repository. I will have to remember to remove the 
changed versions.

omh/faq.omh:
The CppAD FAQ has been brought up to date.


git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@4 0e562018-b8fb-0310-a933-ecf073c7c197