Overall Goal: Make it possible to separately reference the Whats New, Bugs, and Wish List information.
omh/WhatsNew05.omh
The Bug and Wish List information as moved out of this section
and into their own separate sections in the documentation.
In addition, the previous years Whats New section were placed
below the current year.

omh/Bugs.omh and omh/WishList.omh
Theses new sections are part of the appendix
in the documentation.

omh/Appendix.omh
This section was modified to affect the changes mentioned above.


git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@36 0e562018-b8fb-0310-a933-ecf073c7c197