For sections below AD, mention in docs which operations are not taped.
SvnCommit: file that preformed this commit.
WhatsNew06.omh: users view of the changes.
Atan2.h: fix missing parenthesis in documentation syntax.
Value.h: mention that this operation is not taped.
Output.h: mention that this operation is not taped.
Integer.h: mention that this operation is not taped.
Compare.h: mention that these operations are not taped.
----------------------------------------------------------------------
Edit section above for individual cases, section below stays the same
----------------------------------------------------------------------


git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@316 0e562018-b8fb-0310-a933-ecf073c7c197