trunk: Change stable versions to different numbering system (not date).
gpl_license.sh: get version in way that works for versions that are not dates.
dos_format.sh: get version in way that works for versions that are not dates.
build.sh: version (as above), only distribute *.tgz, fix help message.
new_stable.sh: major changes (that started with version 2.3).



git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@1147 0e562018-b8fb-0310-a933-ecf073c7c197